PlataformaSaude/main/java/br/com/furiossam/PlataformaSaude/entity/Atendimento.java
package br.com.furiossam.PlataformaSaude.entity;

import java.util.Date;

public class Atendimento {
	
	private long id;
	private Date dataHora;
	private String nomeProcedimento;
	private String modalidade;
	private Medico medico;
	
	public long getId() {
		return id;
	}
	public void setId(long id) {
		this.id = id;
	}
	public Date getDataHora() {
		return dataHora;
	}
	public void setDataHora(Date dataHora) {
		this.dataHora = dataHora;
	}
	public String getNomeProcedimento() {
		return nomeProcedimento;
	}
	public void setNomeProcedimento(String nomeProcedimento) {
		this.nomeProcedimento = nomeProcedimento;
	}
	public String getModalidade() {
		return modalidade;
	}
	public void setModalidade(String modalidade) {
		this.modalidade = modalidade;
	}
	public Medico getMedico() {
		return medico;
	}
	public void setMedico(Medico medico) {
		this.medico = medico;
	}
	
	
	

}
